First attempt at introducing a SAX-like layer between the parser and the DOM.
- sax-handler.lisp defines the generic functions a handler should implement.
  Default methods are provided that simply do nothing.
  This is currently horribly incomplete.
- dom-builder.lisp implements these methods to build a dom-impl:document
- xml-parser.lisp now call the sax methods instead of the dom constructors.
  It uses the special variable *handler* (formerly known as *document*), which
  comes from a new optional argument to parse-file/stream/string. This argument
  defaults to a dom-impl:dom-builder, so the default behaviour when called
  with just one arg is unchanged.
- Additionally, the parse-* functions are now exported from the XML package.

Even if the sax method signatures imply otherwise, there is no support for
namespaces yet. The namespace-uri and prefix arguments are always nil, the
qname contains the full GI. This should obviously be changed, but it requires
some good idea how to represent attributes in start-element (among other things).
Currently it's just a (name . value) alist, which doesn't quite work with
namespaced attributes.