Since closure-frame-top-level is no longer used, comment it out.
Use a :before method on run-frame-top-level to set
*closure-inited-p*. (this used to be done by closure-frame-top-level)

Reimplement the quit command using frame-exit, since the 'closure-quit
catch tag no longer exists.

Use process-interrupt from the clim-sys package, instead of from the
mp package, since clim-sys should be more portable.

Turn on scrollbars for the interactor pane, since otherwise it trashes
the status line on the bottom of the window and stops updating.