Complete the renaming *MEDIUM* -> *PANE*.
Panes are CLIM extended-streams, and remember output to them in output
records.  Mediums are much simpler, and don't have this kind of
memory.  So, though the same drawing functions (DRAW-TEXT, DRAW-LINE)
can have the same initial effect applied to a pane and a medium, the
output-record state is very different.